pyproject.toml
[tool.poetry]
name = "filament"
version = "0.1.0"
description = ""
authors = ["Pleasant-Cry <pleasant.cry@gmail.com>"]
license = "AGPL-3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
annoy = "^1.17.1"
emoji = "^2.2.0"
fastapi = "^0.95.0"
pandas = "^1.5.3"
sentence-transformers = "^2.2.2"
spacy = "^3.5.1"
torch = "^2.0.0"
torchaudio = "^2.0.1"
torchvision = "^0.15.1"
transformers = "^4.27.4"
uvicorn = "^0.21.1"
httpx = "^0.23.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
ipython = "^8.12.0"
pyflakes = "^3.0.1"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py310']

[tool.isort]
profile = "black"
